+++ helper/basic-auth.ts
@@ -16,7 +16,8 @@
 
 "use strict";
 
-const config = require("./../config")();
+import { Config } from "./../config";
+import { Buffer } from 'buffer';
 
 /**
  * Authorization (basic auth) for dialogflow fulfillment request
@@ -24,7 +25,7 @@ const config = require("./../config")();
  * @param {object} res http response
  * @param {function} next invokes the succeeding middleware/function
  */
-const basicAuth = (req, res, next) => {
+export const basicAuth = (req, res, next) => {
     const auth = req.get("authorization");
     if (req.path === "/healthcheck") {
         next();
@@ -32,13 +33,11 @@ const basicAuth = (req, res, next) => {
     else if (!auth) {
         res.status(401).send({ "status": 401, "message": "Unauthorized" });
     } else {
-        const credentials = new Buffer.from(auth.split(" ").pop(), "base64").toString("ascii").split(":");
-        if (credentials[0] === config.auth.username && credentials[1] === config.auth.password) {
+        const credentials = Buffer.from(auth.split(" ").pop(), "base64").toString("ascii").split(":");
+        if (credentials[0] === Config.auth.username && credentials[1] === Config.auth.password) {
             next();
         } else {
             res.status(401).send({ "status": 401, "message": "Unauthorized" });
         }
     }
 };
-
-module.exports = basicAuth;
