+++ intent-library/intent-mapper.ts
@@ -1,0 +1,26 @@
+/**
+ * Copyright 2020 Quantiphi, Inc. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+"use strict";
+import { defaultWelcomeIntent } from "./intents/default-welcome-intent";
+
+/**
+ * Maps dialogflow intent with its controller
+ */
+
+export const intentMapper = {
+    "Default Welcome Intent": defaultWelcomeIntent,
+};
