+++ package.json
@@ -7,8 +7,9 @@
   "license": "Apache-2.0",
   "main": "server.js",
   "scripts": {
-    "start": "node server.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "build": "tsc server.ts --outDir dist",
+    "start": "npm run build && node dist/server.js",
+    "test": "tsc ./specs/<path_to_test_file>.spec.ts --outDir dist/specs && mocha 'dist/specs/**/*.spec.js'"
   },
   "engines": {
     "node": ">=6.0.0"
@@ -37,5 +38,14 @@
     "boilerplate",
     "dialogflow boilerplate",
     "fulfillment service"
-  ]
+  ],
+  "devDependencies": {
+    "@types/chai": "^4.3.0",
+    "@types/express": "^4.17.13",
+    "@types/mocha": "^9.1.0",
+    "@types/node": "^17.0.21",
+    "chai": "^4.3.6",
+    "mocha": "^9.2.1",
+    "typescript": "^4.6.2"
+  }
 }
